Create greetings component
