+++ app/javascript/components/App.js
@@ -1,31 +1,11 @@
-import React, { useEffect } from 'react';
-import { connect, useDispatch, useSelector } from 'react-redux';
-import { setGreeting } from '../actions/greetingActions';
+import React from 'react';
+import Greetings from './greetings'
 
 function App() {
-  const greeting = useSelector(state => state.greeting.text);
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    fetch('/greeting/api')
-      .then(response => response.json())
-      .then(data => {
-        dispatch(setGreeting(data.text));
-      })
-      .catch(error => {
-        console.error(error);
-      });
-  }, [dispatch]);
 
   return (
-    <h1>
-      {greeting}
-    </h1>
+    <Greetings />
   );
 }
 
-const mapStateToProps = (state) => ({
-  greeting: state.greeting.text,
-});
-
-export default connect(mapStateToProps)(App);
+export default App;
